requirements.txt
flask
werkzeug
flask-cors
pytesseract
requests
pillow
openai
python-dotenv
opencv-python
numpy